requirements.txt
aniso8601==9.0.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==22.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
dnspython==2.2.1
ec2-hibinit-agent==1.0.0
ecdsa==0.18.0
email-validator==1.2.1
Flask==2.1.2
flask-mongoengine==1.0.0
flask-restx==0.5.1
Flask-WTF==1.0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.1
mongoengine==0.24.2
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
passlib==1.7.4
pexpect==4.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.3.0
pymongo==4.2.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.3.0+ubuntu2.1
python-debian===0.1.43ubuntu1
python-jose==3.3.0
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.9
SecretStorage==3.3.1
service-identity==18.1.0
simplejson==3.17.6
six==1.16.0
sos==4.3
ssh-import-id==5.11
systemd-python==234
Twisted==22.1.0
ubuntu-advantage-tools==27.10
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==2.1.2
WTForms==3.0.1
zipp==1.0.0
zope.interface==5.4.0
